+++ Siv3D/src/Siv3D/Gamepad/IGamepad.hpp
@@ -39,9 +39,9 @@ namespace s3d
 
 		virtual bool up(size_t playerIndex, uint32 index) = 0;
 
-		virtual void clearInput(uint32 playerIndex, uint32 index) = 0;
+		virtual void clearInput(size_t playerIndex, uint32 index) = 0;
 
-		virtual bool cleared(uint32 playerIndex, uint32 index) = 0;
+		virtual bool cleared(size_t playerIndex, uint32 index) = 0;
 
 		virtual Duration pressedDuration(size_t playerIndex, uint32 index) = 0;
